WYSIWYG Input Action Positions.
