+++ Assets/MixedRealityToolkit/_Core/Inspectors/Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 01081ff15c3de9a45834acf3cbf52c00
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
