+++ Assets/MixedRealityToolkit/_Core/Inspectors/Data/EditorWindowOptions.json
@@ -1,0 +1,1 @@
+{"Controllers":[{"Controller":6,"Handedness":2,"InputLabelPositions":[{"x":296.0,"y":69.0},{"x":161.0,"y":300.0},{"x":162.0,"y":279.0},{"x":483.0,"y":202.0},{"x":481.0,"y":256.0},{"x":482.0,"y":230.0},{"x":63.0,"y":176.0},{"x":91.0,"y":208.0},{"x":485.0,"y":133.0},{"x":483.0,"y":154.0},{"x":481.0,"y":179.0},{"x":108.0,"y":243.0}]},{"Controller":8,"Handedness":0,"InputLabelPositions":[{"x":185.0,"y":392.0},{"x":181.0,"y":412.0},{"x":439.0,"y":401.0},{"x":438.0,"y":424.0},{"x":42.0,"y":293.0},{"x":288.0,"y":49.0},{"x":78.0,"y":116.0},{"x":534.0,"y":116.0},{"x":248.0,"y":107.0},{"x":371.0,"y":108.0},{"x":63.0,"y":138.0},{"x":545.0,"y":140.0},{"x":547.0,"y":298.0},{"x":571.0,"y":278.0},{"x":577.0,"y":254.0},{"x":564.0,"y":233.0}]},{"Controller":6,"Handedness":1,"InputLabelPositions":[{"x":299.0,"y":80.0},{"x":400.0,"y":339.0},{"x":462.0,"y":282.0},{"x":94.0,"y":198.0},{"x":92.0,"y":216.0},{"x":92.0,"y":236.0},{"x":499.0,"y":203.0},{"x":500.0,"y":226.0},{"x":499.0,"y":140.0},{"x":499.0,"y":161.0},{"x":498.0,"y":182.0},{"x":489.0,"y":249.0}]},{"Controller":4,"Handedness":1,"InputLabelPositions":[{"x":556.0,"y":156.0},{"x":27.0,"y":95.0},{"x":22.0,"y":145.0},{"x":26.0,"y":122.0},{"x":58.0,"y":65.0},{"x":129.0,"y":355.0},{"x":138.0,"y":35.0},{"x":511.0,"y":74.0},{"x":523.0,"y":98.0},{"x":530.0,"y":119.0},{"x":12.0,"y":253.0},{"x":10.0,"y":294.0},{"x":42.0,"y":203.0},{"x":10.0,"y":273.0},{"x":14.0,"y":315.0},{"x":477.0,"y":28.0},{"x":503.0,"y":53.0}]},{"Controller":4,"Handedness":2,"InputLabelPositions":[{"x":32.0,"y":169.0},{"x":5.0,"y":242.0},{"x":6.0,"y":260.0},{"x":7.0,"y":280.0},{"x":7.0,"y":298.0},{"x":428.0,"y":384.0},{"x":429.0,"y":114.0},{"x":398.0,"y":93.0},{"x":360.0,"y":71.0},{"x":325.0,"y":51.0},{"x":495.0,"y":154.0},{"x":496.0,"y":203.0},{"x":494.0,"y":176.0},{"x":495.0,"y":225.0},{"x":305.0,"y":28.0},{"x":85.0,"y":29.0}]}]}
