+++ Assets/MixedRealityToolkit/_Core/Inspectors/Data/EditorWindowOptions.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 28091d5ea9b5739419a221a06fa1ec89
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
